Crazy config decision
