--- .vscode/.settings.json
@@ -1,8 +1,0 @@
-{
-  "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
-  },
-  "javascript.validate.enable": false,
-  "js/ts.implicitProjectConfig.checkJs": true,
-  "emmet.triggerExpansionOnTab": true
-}
