--- .vscode/workspace.code-workspace
@@ -1,13 +1,0 @@
-{
-	"folders": [
-		{
-			"path": ".."
-		},
-		{
-			"path": "../src"
-		}
-	],
-	"settings": {
-		"typescript.tsdk": "gamma/node_modules/typescript/lib"
-	}
-}
