+++ babel.config.js
@@ -1,19 +1,1 @@
-module.exports = function (api) {
-  api.cache(true);
-
-  const plugins = [
-    // Absolute module imports (to work when app is built)
-    [
-      'module-resolver',
-      {
-        root: ['src'],
-        extensions: ['.ts', '.tsx', '.json'],
-      },
-    ],
-  ];
-
-  return {
-    presets: ['babel-preset-expo'],
-    plugins,
-  };
-};
+module.exports = require('./config/babel.config.js');
