+++ jest.config.js
@@ -1,4 +1,1 @@
-module.exports = {
-  preset: 'react-native',
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-};
+module.exports = require('./config/jest.config.js');
