+++ tsconfig.json
@@ -1,23 +1,3 @@
 {
-  "extends": "expo/tsconfig.base",
-  "compilerOptions": {
-
-    // Absolute module imports (for vscode intellisense)
-    "baseUrl": "src",
-
-    "strict": true,
-    "plugins": [
-
-      // For rare properties to be valid in styled CSS
-      {
-        "name": "typescript-styled-plugin",
-        "lint": {
-          "validProperties": [
-            "elevation"
-          ]
-        }
-      }
-
-    ]
-  },
+  "extends": "./config/tsconfig.json",
 }
